requirements.txt
backcall==0.2.0
cycler==0.10.0
Cython==0.29.24
decorator==5.1.0
ipython==7.28.0
jedi==0.18.0
joblib==1.1.0
kiwisolver==1.3.2
linearmodels==4.24
matplotlib==3.4.3
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.21.3
pandas==1.3.4
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.21
property-cached==1.6.4
ptyprocess==0.7.0
pyarrow==5.0.0
Pygments==2.10.0
pyhdfe==0.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
statsmodels==0.13.0
tqdm==4.62.3
traitlets==5.1.0
wcwidth==0.2.5
